Code after lesson 13.
There seems to be a bug where the number of accesses reported on the first line of the log file can be higher than the number of entires in the log file.  It has only been observed to be one higher, and it appears that the first access is the one that is lost.

It is suspected that this happens while the file is loading.

Since this is the last piece of the tutorial, it's unlikely that it will be investigated further.