Code after lesson 6.
In this lesson the instructor demonstrates static files by telling the viewer to copy and paste all of the bootstrap css bundle form a CDN.  It has the same license as the one I am using for this project (MIT) but I'm not entirely comfortable doing that.

Then the instructor wants to demonstrate a static image, so they use Google image search for pictures of puppies and use one of the front page results.  That's not very respectful of copyright in a video that is being sold for profit.

My thoughts are that the instructor should try just a little bit more.