Pythonmorsels cached_property passing all objectives.
This has been updated to reflect suggestions in the reference solution.  In particular, in **__set__** the cached value is cleared after the setter is called.

Also, it isn't necessary to return a new **cached_property** object from **cached_property.setter** and the like, as long as the modified object returns itself.